app_orcamento_standalone/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
py7zr
pandas
openpyxl
xlrd
email-validator
slowapi
httpx
redis
aioredis
requests
